only save/delete metadata if it is posted

+++ admin/level-edit.php
@@ -438,13 +438,15 @@ final class RUA_Level_Edit {
 
 		// Update metadata
 		foreach (RUA_App::instance()->level_manager->metadata()->get_all() as $field) {
-			$new = isset($_POST[$field->get_id()]) ? $_POST[$field->get_id()] : '';
+			$new = isset($_POST[$field->get_id()]) ? $_POST[$field->get_id()] : false;
 			$old = $field->get_data($post_id);
 
-			if ($new != '' && $new != $old) {
-				$field->update($post_id,$new);
-			} elseif ($new == '' && $old != '') {
-				$field->delete($post_id,$old);
+			if($new !== false) {
+				if ($new != '' && $new != $old) {
+					$field->update($post_id,$new);
+				} elseif ($new == '' && $old != '') {
+					$field->delete($post_id,$old);
+				}
 			}
 		}
 	}
